requirements.txt
requests==2.26.0
python-dotenv==0.19.2
terminaltables==3.1.0
